refactor: Factor out ability to add custom validators/casters for php types
PHP types are standardized in how they work. A validator or caster, while being able to modify the implementation, would never have their behaviour changed. Its better for the library or a third-party to do this so there are no incompatibilities between how php handles it and how a custom validator/caster might handle it.